backend/requirements.txt
flask
flask-cors
snowflake-connector-python
python-dotenv
